Pin whitenoise dependencies to 3.3.1
Version 4.0 introduces breaking changes, and the wazimap framework will
need to be upgraded to handle that change. There is a pull request open
for that. See https://github.com/OpenUpSA/wazimap/pull/133

Until that is merged (and nepalmap is upgraded to a version that
includes that change, this pins the whitenoise dependency to 3.3.1.